android: update to ndk15b - bump min target api to 14
