+++ code/demo/android-native/app/build.gradle
@@ -1,15 +1,20 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 19
-    buildToolsVersion "20.0.0"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.3"
 
     defaultConfig {
         applicationId "org.orxproject.orxtestnative"
-        minSdkVersion 10
+        minSdkVersion 14
         targetSdkVersion 19
         versionCode 1
         versionName "1.0"
+
+        ndk{
+            abiFilters "x86", "armeabi-v7a", "arm64-v8a"
+        }
+
     }
     buildTypes {
         release {
@@ -18,8 +23,10 @@ android {
         }
     }
 
-    sourceSets.main {
-        jniLibs.srcDirs = ['src/libs']
+    externalNativeBuild {
+        ndkBuild {
+            path 'src/main/cpp/Android.mk'
+        }
     }
 }
