+++ code/demo/android-native/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
              This will take care of integrating with our NDK code. -->
         <activity android:name=".MainActivity"
             android:label="@string/app_name"
-            android:screenOrientation="@integer/orientation"
+            android:screenOrientation="sensorPortrait"
             android:configChanges="orientation|keyboardHidden|screenSize">
 
             <!-- Tell NativeActivity the name of our .so -->
