--- code/demo/android/app/src/jni/Application.mk
@@ -1,4 +1,0 @@
-APP_ABI := armeabi-v7a x86 arm64-v8a
-APP_MODULES = orxTest
-NDK_TOOLCHAIN_VERSION := clang
-APP_PLATFORM = android-9
