+++ code/demo/android/app/src/main/cpp/Application.mk
@@ -1,0 +1,4 @@
+APP_ABI := armeabi-v7a x86 arm64-v8a
+APP_MODULES = orxTest
+NDK_TOOLCHAIN_VERSION := clang
+APP_PLATFORM = android-14
