--- code/demo/android/app/src/main/res/values-television/config.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <integer name="orientation">0</integer>
-</resources>
