--- code/demo/android/app/src/main/res/values-v11/themes.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <style name="MyTheme" parent="@android:style/Theme.Holo.NoActionBar.Fullscreen"/>
-</resources>
